system/yolo_backend/requirements.txt
# YOLOv11 數位雙生分析系統依賴套件

# 核心 AI/ML 框架
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0

# Web 框架與 API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 資料處理與分析
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# 資料處理與驗證
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# 基礎工具
requests>=2.31.0
aiofiles>=23.2.0

# 系統監控
psutil>=5.9.0

# 日誌處理
structlog>=23.2.0
python-json-logger>=2.0.0

# 開發與測試工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# 資料庫套件
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
